LayoutTests/fast/canvas/translate-text-expected.txt

Test for chromium's canvas bug where fillText resets the current context https://bugs.webkit.org/show_bug.cgi?id=26436 . You must see green box with message 'This text should be seen', and black box below the green box.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pixel 0, 0 is [0, 255, 0, 255]
PASS successfullyParsed is true

TEST COMPLETE
